# Faster TrinaryLogic->and()