@kamil-tekiela 2025-03-08
Does it make any sense for the `or` method to be aligned with this change?