@staabm 2025-03-09
I did not yet see a profile in which or() was slow, but we could for consistency